Stable Bluetooth Transfer
This commit marks the first round of stable bluetooth transfers. The process is relatively automated, but there are a couple of tweaks that had to be made:

The browser now has a new invitation pending state to match the advertiser (this state needs to send an alert to the user asking for confirmation -- it still accepts by default). Upon session disconnections at the end of a transfer, the MatchTransfer session is destroyed and a new one is created. This removes all previous knowledge of the connections from a previous transfer to help make things run smoother. 
Various member level variables are reset upon transfer completion such as the list of found peers, the session state. All delegates are notified of the reset state as well so they can act accordingly.

Issue: #7