Start implementing match data transfer
This commit adds the ability to transfer match data instead of just pings.  The data is encoded to json, then sent as data. The matches then are captured and need to be added to the browsers match store. 

This commit also changes the mock match store to be enabled by a static var instead of a new static store instance. If the useMock var is set before the first initialization, the shared store will be a mock store instead of looking at the archive. 

Issue: #7