Add Data Transfer Import
This commit adds the final step in data transfer - importing matches! Testing needs to be done, but this should add matches. The regression that needs to be addressed is dealing with matches imported that are the same as matches on file (same team number and match). This can be addressed in a follow up update.

Issue: #7