Fix TeamInfo No Show
This commit fixes an issue when trying to save a no show. Because the next button was mapped to certain functions and was also mapped directly to a segue in the interface builder, the no show segue would be ignored. The segue to data entry has been moved to a code triggered location and now the nextButtonTap function will perform as expected.

Issue: #4