Fix Match Encoding Bug
This commit fixes an issue with initializing matches with the MatchEncodingHelper. The matches by default are initialized as MatchImpl, which doesn't work if the matches saved are subclasses. This means that game specific match data is lost from closing and reopening the app.

To fix this, the match type is saved as a string along with the rest of the match data. When initalizing from the archive, the match type is read and that type initializer is called instead of the generic one. This allows the MatchEncodingHelper to initialize the right class dynamically.

Issue: #9