Add PowerScoutDebugging Target
This commit adds a PowerScoutDebugging target that is a separate app from the main release build. With this new app, a launchpad exists to launch targetted views without having to navigate through the entire app. This target can be used to test out and debug certain views instead of running the main app. This will help isolate problems to a specific view.

The debug launchpad currently contains a segue to the Data Entry view and to the debug data transfer view.

Issue: #11