Correctly Initialize Matches During Import
This commit fixes a bug associated with importing data transferred through bluetooth. The match data was using to the default MatchImpl class instead of specific subclasses. Instead of doing this for all cases. the matchType key was checked to make sure a valid matchType was given along with the data. This match type was then used as the initializer in order to get the correct subclass.

Issue: #53