Fix sendData bug
This commit fixes a bug that could occur when using CB transfers instead of MP. The sendData function runs at the correct time in all cases, but this is only required for MP transfers. During a CB transfer, this don't have a lasting effects, but does produce some red herring error messages. 

To fix this, a check was added to make sure we are performing a MP transfer before trying to send data.

Issue: #67