Add git tag script
This commit changes the version_bump.sh script with an xcode prefix
so a git_version_bump.sh script could be made. This script calls the
xcode version bump script, forwarding the type passed along. It also
automatically commits the changes (under a bot name for consistency)
and tags the new version so the xcode and git version are kept in
sync.

Issues: #73