Add Branch Check
This commit changes the git version bump script to make sure version bumps only occur on the master branch. This is a change to ensure that tagging only happens on the master branch.

In the future, this will be replaced by a bot that is capable of creating a separate branch for the xcode plist change, then on a merge, it will tag the master branch.

Issue: #73